SpringApp * サインアップ機能微調整。

+++ common/src/main/java/cherry/spring/common/helper/signup/SignupRequestHelper.java
@@ -20,7 +20,7 @@ import org.joda.time.LocalDateTime;
 
 public interface SignupRequestHelper {
 
-	int createSignupRequest(String mailAddr, String token,
+	long createSignupRequest(String mailAddr, String token,
 			LocalDateTime appliedAt);
 
 	boolean validateMailAddr(String mailAddr, LocalDateTime intervalFrom,
